feat: add caching feature
A basic cachesim feature has been added. It stores newly searched words
in a file named `.cambd-cache.json` in `$HOME`. And for every new word
lookup it checks the cache file if the definition for that specfic word
already exists. If yes then return immendietly else fetch the new
definition and add in the cache file for future reference.

fixes #6