feat: add caching feature (#7)
A basic caching feature has been added. It stores newly searched words
in a file named .cambd-cache.json in $HOME. And for every new word
lookup, it checks the cache file if the definition for that specific word
already exists. If yes then return immediately else fetch the new
definition and add it to the cache file for future reference.

fixes #6